fix: change IncrementFailedAttempts return type from int64 to int
DB の RETURNING 値を int64 でスキャンしていたが、model.Authentication および
model.Admin の FailedAttempts フィールドは int 型のためモデルと不一致だった。
インタフェース定義・実装の両方を int に統一する。

+++ backend/internal/domain/repository/admin.go
@@ -14,7 +14,7 @@ type AdminRepository interface {
 	Create(ctx context.Context, admin *model.Admin) error
 	Count(ctx context.Context) (int, error)
 	UpdatePassword(ctx context.Context, id int, passwordHash string) error
-	IncrementFailedAttempts(ctx context.Context, id int) (int64, error)
+	IncrementFailedAttempts(ctx context.Context, id int) (int, error)
 	LockAccount(ctx context.Context, id int, until time.Time) error
 	UpdateLoginSuccess(ctx context.Context, id int) error
 }
